requirements.txt
anthropic==0.18.1
python-dotenv==1.0.0
opencv-python==4.8.1.78